requirements.txt
requests
python-dotenv
faiss-cpu 
sentence-transformers